Moves each component into its own folder
