+++ gui/src/components/Graphs/Graph/Graph.css
