+++ gui/src/components/Graphs/GraphSelector/GraphSelector.css
