+++ gui/src/components/Indicators/GenericIndicator/GenericIndicator.css
