+++ gui/src/components/Indicators/IndicatorTable/IndicatorTable.css
