+++ gui/src/components/MissionStatus/CurrentStatus.css
